feat: hide headers if they are hidden by logic
