+++ frontend/src/features/public-form/PublicFormProvider.tsx
@@ -14,7 +14,6 @@ import { isEqual } from 'lodash'
 import get from 'lodash/get'
 import simplur from 'simplur'
 
-import { BasicField } from '~shared/types'
 import {
   FormAuthType,
   FormResponseMode,
@@ -42,11 +41,7 @@ import {
 
 import { FormNotFound } from './components/FormNotFound'
 import { usePublicAuthMutations, usePublicFormMutations } from './mutations'
-import {
-  PublicFormContext,
-  SidebarSectionMeta,
-  SubmissionData,
-} from './PublicFormContext'
+import { PublicFormContext, SubmissionData } from './PublicFormContext'
 import { usePublicFormView } from './queries'
 
 interface PublicFormProviderProps {
@@ -290,27 +285,6 @@ export const PublicFormProvider = ({
     [cachedDto?.form, cachedDto?.spcpSession],
   )
 
-  const sectionScrollData = useMemo(() => {
-    const { form } = cachedDto ?? {}
-    if (!form || isAuthRequired) {
-      return []
-    }
-    const sections: SidebarSectionMeta[] = []
-    if (form.startPage.paragraph)
-      sections.push({
-        title: 'Instructions',
-        _id: 'instructions',
-      })
-    form.form_fields.forEach((f) => {
-      if (f.fieldType !== BasicField.Section) return
-      sections.push({
-        title: f.title,
-        _id: f._id,
-      })
-    })
-    return sections
-  }, [cachedDto, isAuthRequired])
-
   if (isNotFormId) {
     return <NotFoundErrorPage />
   }
@@ -323,7 +297,6 @@ export const PublicFormProvider = ({
         formId,
         error,
         submissionData,
-        sectionScrollData,
         isAuthRequired,
         captchaContainerId: containerId,
         expiryInMs,
